revert results changes
