+++ aeon/visualisation/results/tests/test_results_plotting.py
@@ -1,4 +1,4 @@
-"""Tests for the scatter diagram makers."""
+"""Tests for the critical difference diagram maker."""
 
 import os
 
@@ -6,15 +6,46 @@ import numpy as np
 import pytest
 
 import aeon
-from aeon.benchmarking import get_estimator_results_as_array
+from aeon.benchmarking.results_loaders import get_estimator_results_as_array
 from aeon.datasets.tsc_data_lists import univariate_equal_length
 from aeon.utils.validation._dependencies import _check_soft_dependencies
-from aeon.visualisation import plot_pairwise_scatter, plot_scatter_predictions
+from aeon.visualisation import (
+    plot_boxplot_median,
+    plot_scatter,
+    plot_scatter_predictions,
+)
 
 data_path = os.path.join(
     os.path.dirname(aeon.__file__),
     "benchmarking/example_results/",
 )
+
+
+@pytest.mark.skipif(
+    not _check_soft_dependencies("matplotlib", severity="none"),
+    reason="skip test if required soft dependency not available",
+)
+def test_plot_boxplot_median():
+    from matplotlib.figure import Figure
+
+    cls = ["HC2", "FreshPRINCE", "InceptionT", "WEASEL-D"]
+
+    data = univariate_equal_length
+    res = get_estimator_results_as_array(
+        estimators=cls, datasets=data, path=data_path, include_missing=True
+    )
+
+    fig = plot_boxplot_median(res, cls, plot_type="violin")
+    assert isinstance(fig, Figure)
+    fig = plot_boxplot_median(res, cls, plot_type="boxplot", outliers=False)
+    assert isinstance(fig, Figure)
+    fig = plot_boxplot_median(res, cls, plot_type="swarm")
+    assert isinstance(fig, Figure)
+    fig = plot_boxplot_median(res, cls, plot_type="strip")
+    assert isinstance(fig, Figure)
+
+    with pytest.raises(ValueError):
+        plot_boxplot_median(res, cls, plot_type="error")
 
 
 @pytest.mark.skipif(
@@ -49,9 +80,7 @@ def test_plot_scatter():
     res = get_estimator_results_as_array(
         estimators=cls, datasets=data, path=data_path, include_missing=True
     )
-    fig = plot_pairwise_scatter(
-        res, cls[0], cls[1], metric="accuracy", statistic_tests=True
-    )
+    fig = plot_scatter(res, cls[0], cls[1], metric="accuracy", statistic_tests=True)
     assert isinstance(fig, Figure)
 
     cls = ["InceptionTime", "WEASEL-D"]
@@ -60,7 +89,7 @@ def test_plot_scatter():
     res = get_estimator_results_as_array(
         estimators=cls, datasets=data, path=data_path, include_missing=True
     )
-    fig = plot_pairwise_scatter(res, cls[0], cls[1], metric="accuracy", title="test")
+    fig = plot_scatter(res, cls[0], cls[1], metric="accuracy", title="test")
     assert isinstance(fig, Figure)
 
     cls = ["InceptionTime", "WEASEL-D"]
@@ -69,7 +98,7 @@ def test_plot_scatter():
     res = get_estimator_results_as_array(
         estimators=cls, datasets=data, path=data_path, include_missing=True
     )
-    fig = plot_pairwise_scatter(
+    fig = plot_scatter(
         1 - res,
         cls[0],
         cls[1],
@@ -79,7 +108,7 @@ def test_plot_scatter():
     )
     assert isinstance(fig, Figure)
 
-    fig = plot_pairwise_scatter(
+    fig = plot_scatter(
         1 - res,
         cls[0],
         cls[1],
@@ -90,6 +119,11 @@ def test_plot_scatter():
     assert isinstance(fig, Figure)
 
     with pytest.raises(ValueError):
-        plot_pairwise_scatter(res, cls[0], cls[1], metric="error", lower_better=False)
+        plot_scatter(res, cls[0], cls[1], metric="error", lower_better=False)
     with pytest.raises(ValueError):
-        plot_pairwise_scatter(res, cls[0], cls[1], metric="accuracy", lower_better=True)
+        plot_scatter(res, cls[0], cls[1], metric="accuracy", lower_better=True)
+
+
+def test_plot_multi_comparison_matrix():
+    # to be completed when MCM is implemented
+    pass
